Draft: Revisits strict build setting
- Strict build for all projects including tests
  - To reduce bug in test code
- Introduce option to disable warning for external libraries
  - To reduce warning message from external libraries

ONE-DCO-1.0-Signed-off-by: Hyeongseok Oh <hseok82.oh@samsung.com>

+++ compute/cker/src/train/Adam.test.cc
@@ -72,7 +72,7 @@ private:
 
     const T alpha = _learning_rate * std::sqrt(static_cast<T>(1) - beta2_power) /
                     (static_cast<T>(1) - beta1_power);
-    for (int i = 0; i < _expected.size(); ++i)
+    for (size_t i = 0; i < _expected.size(); ++i)
     {
       T m = _m.at(i);
       T v = _v.at(i);
